Encrypt and share any object

Encrypt and share R objects on github

hugo_share_secret(object, passphrase)

hugo_get_secret(user, repo_name, object_name, passphrase)

Arguments

object
object to share

passphrase
object will be encrypted using provided passphrase

user
Github user name, to construct proper url to repository.

repo_name
Name of github repository, also used to construct url.

object_name
Name of the object you want to download.

Value

Please check Functions section.

Functions

hugo_share_secret: Shares given object via github, prompts user for input.

Returns: in case of success prompts user with message, and returns string, which is direct call to hugo_get_secret. In case of failure, prompts user with appropriate message and returns NULL.

hugo_get_secret: Downloads encrypted object from github. Returns: in case of success prompts user with message, and returns object. It us up to user how to handle it - wheter to assign it to variable, or use as parameter. In case of failure, prompts user with appropriate message and returns NULL.

Examples

# NOT RUN {
hugo_share_secret(iris, "secret")
# }